refactor
https://github.com/metasfresh/metasfresh/issues/5266
